Fix windows build
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@41552 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/common/filesystem/filesystemext.cpp
@@ -231,7 +231,7 @@ static int rtFsExtLoadBlkGrpDesc(PRTFILESYSTEMEXT pThis, uint32_t iBlkGrp)
 }
 
 static bool rtFsExtIsBlockRangeInUse(PRTFILESYSTEMEXTBLKGRP pBlkGrpDesc,
-                                     uint32_t offBlockStart, uint32_t cBlocks)
+                                     uint32_t offBlockStart, size_t cBlocks)
 {
     bool fUsed = false;
